태스크 권한 설정

+++ app/Http/Controllers/TaskController.php
@@ -9,7 +9,7 @@ class TaskController extends Controller
 {
     public function index()
     {
-        $tasks = Task::latest()->get();
+        $tasks = Task::latest()->where('user_id', auth()->id())->get();
 
         return view('tasks.index', [
             'tasks' => $tasks
@@ -28,18 +28,29 @@ class TaskController extends Controller
             'body'=>'required'
         ]);
 
-        $task = Task::create(request(['title', 'body'])); //이름이 같으면 매칭된다.
+        $values = request(['title', 'body']);
+        $values['user_id'] = auth()->id();
+
+        $task = Task::create($values); //이름이 같으면 매칭된다.
         
         // $tast = Task::create([
         //     'title' => request->input('title'),
         //     'body' => request->input('body')
         // ]);
 
-        return redirect('/tasks/'.$tast->id);
+        return redirect('/tasks/'.$task->id);
     }
 
     public function show(Task $task)
     {
+        // if(auth()->id() != $task->user_id) {
+        //     abort(403);
+        // }
+
+        // abort_if(!auth()->user()->owns($task), 403); 모두 같은 방법
+        // abort_unless(auth()->user()->owns($task), 403);
+        abort_if(auth()->id() != $task->user_id, 403); //축약
+
         return view('tasks.show', [
             'task' => $task
         ]);
@@ -47,6 +58,8 @@ class TaskController extends Controller
 
     public function edit(Task $task)
     {
+        abort_if(auth()->id() != $task->user_id, 403);
+
         return view('tasks.edit', [
             'task' => $task
         ]);
@@ -54,6 +67,8 @@ class TaskController extends Controller
 
     public function update(Task $task)
     {
+        abort_if(auth()->id() != $task->user_id, 403);
+        
         request()->validate([
             'title' => 'required',
             'body' => 'required'
@@ -70,6 +85,8 @@ class TaskController extends Controller
 
     public function destroy(Task $task)
     {
+        abort_if(auth()->id() != $task->user_id, 403);
+        
         $task->delete();
         return redirect('/tasks');
     }
